move preview component to new lecture directory
