+++ web-ui/src/components/lecture/preview.less
