+++ web-ui/src/components/lecture/preview.tsx
