+++ web-ui/src/components/questions/questions.tsx
@@ -10,7 +10,7 @@ import { useMutation } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 import apiClient, { ServerErrorResponse, ServerResponse } from '@/http';
 import { history } from 'umi';
-import Preview from '../preview';
+import Preview from '@/components/lecture/preview';
 import { Lecture } from '@/components/lecture';
 import {
   emitEvent,
