+++ web-ui/src/components/selector/search-tool.tsx
@@ -5,7 +5,7 @@ import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
 import apiClient, { ServerErrorResponse, ServerResponse } from '@/http';
 import { history } from 'umi';
-import { PreviewCompact } from '../preview';
+import { PreviewCompact } from '@/components/lecture/preview';
 import {
   emitEvent,
   EVENT_ERROR_RESPONSE,
